sunfish_defconfig: Enable Power Efficient Workqueues
Workqueues marked with WQ_POWER_EFFICIENT are per-cpu by default
but become unbound if workqueue.power_efficient kernel param is
specified.  Per-cpu workqueues which are identified to
contribute significantly to power-consumption are identified and
marked with this flag and enabling the power_efficient mode
leads to noticeable power saving at the cost of smal
performance disadvantage.

Signed-off-by: Sourajit Karmakar <sourajit.karmakar@gmail.com>